+++ CHANGELOG.md
@@ -22,6 +22,9 @@
 * Add hotlinks to the Import log summary for MARC imports (UIDATIMP-988)
 * Cover `<RecentJobLogs>` component with unit tests (UIDATIMP-720)
 * increment stripes to v7 (UIDATIMP-983)
+* Cover `<RepeatableActionsField>` with unit tests (UIDATIMP-722)
+* Cover `<withReferenceValues>` component with unit tests (UIDATIMP-732)
+* Cover `<TreeView>` component with tests (UIDATIMP-728)
 * Cover `<RecordTypesSelect>` component with unit tests (UIDATIMP-721)
 * Validate action profile has a field mapping profile before attaching to job profile (UIDATIMP-990)